+++ prover/src/lib.rs
@@ -16,7 +16,7 @@ pub mod latest_validator_set;
 pub mod merkle_proof;
 pub mod message_sent;
 pub mod next_validator_set;
-pub mod storage_proof;
+pub mod storage_inclusion;
 pub mod validator_set_hash;
 
 pub use common::ProofWithCircuitData;
