+++ prover/src/storage_inclusion/generic_hasher.rs
