+++ prover/src/storage_inclusion/storage_trie_proof/hashed_leaf_parser.rs
