+++ prover/src/storage_inclusion/storage_trie_proof/node_parser/branch_parser/child_node_array_parser/child_node_parser.rs
@@ -16,7 +16,7 @@ use crate::{
         consts::{BLAKE2_DIGEST_SIZE, BLAKE2_DIGEST_SIZE_IN_BITS},
         *,
     },
-    storage_proof::storage_trie_proof::node_parser::{
+    storage_inclusion::storage_trie_proof::node_parser::{
         BranchNodeDataPaddedTarget, MAX_BRANCH_NODE_DATA_LENGTH_IN_BLOCKS, NODE_DATA_BLOCK_BYTES,
     },
     ProofWithCircuitData,
@@ -156,7 +156,7 @@ mod tests {
     use super::{tests_common::*, *};
     use crate::{
         common::array_to_bits,
-        storage_proof::storage_trie_proof::node_parser::compose_padded_node_data,
+        storage_inclusion::storage_trie_proof::node_parser::compose_padded_node_data,
     };
 
     #[test]
@@ -243,7 +243,7 @@ pub mod tests_common {
     use parity_scale_codec::{Compact, Encode};
 
     use super::*;
-    use crate::storage_proof::storage_trie_proof::node_parser::compose_padded_node_data;
+    use crate::storage_inclusion::storage_trie_proof::node_parser::compose_padded_node_data;
 
     #[derive(Clone, Copy)]
     pub enum MockChildType {
